refactor(diff-viewer): callback 配線を分離
